Prodaja nasljeduje Oglas
Napravio da Prodaja nasljeduje Oglas i u Oglas atributu ID promijenio tip iz int u Long jer je ID u Prodaja bio tipa Long.

+++ IzvorniKod/progi-project/src/main/java/progi_project/controller/TicketController.java
@@ -1,12 +1,20 @@
 package progi_project.controller;
 
-import progi_project.model.Ticket;
-import progi_project.service.TicketService;
+import java.util.List;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 
-import java.util.List;
+import progi_project.model.Ticket;
+import progi_project.service.TicketService;
 
 @RestController
 @RequestMapping("/tickets")
@@ -21,7 +29,7 @@ public class TicketController {
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<Ticket> getTicketById(@PathVariable int id) {
+    public ResponseEntity<Ticket> getTicketById(@PathVariable Long id) {
         return ticketService.findById(id)
                 .map(ResponseEntity::ok)
                 .orElse(ResponseEntity.notFound().build());
@@ -33,14 +41,14 @@ public class TicketController {
     }
 
     @PutMapping("/{id}")
-    public ResponseEntity<Ticket> updateTicket(@PathVariable int id, @RequestBody Ticket ticket) {
+    public ResponseEntity<Ticket> updateTicket(@PathVariable Long id, @RequestBody Ticket ticket) {
         ticket.setId(id);
         Ticket updatedTicket = ticketService.updateTicket(ticket);
         return ResponseEntity.ok(updatedTicket);
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteTicket(@PathVariable int id) {
+    public ResponseEntity<Void> deleteTicket(@PathVariable Long id) {
         ticketService.deleteTicket(id);
         return ResponseEntity.noContent().build();
     }
